signup login logout user -> done

+++ accounts/models.py
@@ -8,7 +8,7 @@ import uuid
 class NewUser(AbstractUser):
     id = models.CharField(default=uuid.uuid4, editable=False, primary_key=True)
     phone_number = models.CharField(verbose_name='phone number', null=True, blank=True, unique=True)
-    otp = models.SmallIntegerField(verbose_name='otp code', null=True, blank=True)
+    otp = models.PositiveBigIntegerField(verbose_name='otp code', null=True, blank=True)
     otp_expire = models.DateTimeField(blank=True, null=True, verbose_name='otp code expire time')
     avatar = models.ImageField(upload_to='profile_images', null=True, blank=True, verbose_name='profile image')
     reservation_count = models.SmallIntegerField(default=0, verbose_name='user reservations count')
